Refactor leaderboard services.
* Separated code from LeaderboardsClient into their respective services.
  * LeaderboardsClient contained seemingly-related but ultimately distinct pipelines. This forced the pipelines to supply dependencies that they didn't even use. The separation ensures that the pipelines only require exactly the dependencies that they need.
* Separated code from WorkerRoleBase into their respective services.
  * Like LeaderboardsClient, WorkerRoleBase contained distinct pipelines, mostly due to supplying dependencies for LeaderboardsClient.
* Removed the leaderboard services installer project.
  * This will be replaced by individual installers for each service. This will allow leaderboard services to be updated and managed independently.
