--- toofz.NecroDancer.Leaderboards.Services.Common/ApplicationInsights.config
@@ -1,70 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ApplicationInsights xmlns="http://schemas.microsoft.com/ApplicationInsights/2013/Settings">
-    <TelemetryProcessors>
-        <Add Type="Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse.QuickPulseTelemetryProcessor, Microsoft.AI.PerfCounterCollector"/>
-        <Add Type="Microsoft.ApplicationInsights.Extensibility.AutocollectedMetricsExtractor, Microsoft.ApplicationInsights"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.AdaptiveSamplingTelemetryProcessor, Microsoft.AI.ServerTelemetryChannel">
-            <MaxTelemetryItemsPerSecond>5</MaxTelemetryItemsPerSecond>
-            <ExcludedTypes>Event</ExcludedTypes>
-        </Add>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.AdaptiveSamplingTelemetryProcessor, Microsoft.AI.ServerTelemetryChannel">
-            <MaxTelemetryItemsPerSecond>5</MaxTelemetryItemsPerSecond>
-            <IncludedTypes>Event</IncludedTypes>
-        </Add>
-    </TelemetryProcessors>
-    <TelemetryInitializers>
-        <Add Type="Microsoft.ApplicationInsights.DependencyCollector.HttpDependenciesParsingTelemetryInitializer, Microsoft.AI.DependencyCollector"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.AzureRoleEnvironmentTelemetryInitializer, Microsoft.AI.WindowsServer"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.AzureWebAppRoleEnvironmentTelemetryInitializer, Microsoft.AI.WindowsServer"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.BuildInfoConfigComponentVersionTelemetryInitializer, Microsoft.AI.WindowsServer"/>
-    </TelemetryInitializers>
-    <TelemetryModules>
-        <Add Type="Microsoft.ApplicationInsights.DependencyCollector.DependencyTrackingTelemetryModule, Microsoft.AI.DependencyCollector">
-            <ExcludeComponentCorrelationHttpHeadersOnDomains>
-                <!-- 
-                    Requests to the following hostnames will not be modified by adding correlation headers. 
-                    This is only applicable if Profiler is installed via either StatusMonitor or Azure Extension.
-                    Add entries here to exclude additional hostnames.
-                    NOTE: this configuration will be lost upon NuGet upgrade.
-                -->
-                <Add>core.windows.net</Add>
-                <Add>core.chinacloudapi.cn</Add>
-                <Add>core.cloudapi.de</Add>
-                <Add>core.usgovcloudapi.net</Add>
-                <Add>localhost</Add>
-                <Add>127.0.0.1</Add>
-            </ExcludeComponentCorrelationHttpHeadersOnDomains>
-        </Add>
-        <Add Type="Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.PerformanceCollectorModule, Microsoft.AI.PerfCounterCollector">
-            <!--
-                Use the following syntax here to collect additional performance counters:
-      
-                <Counters>
-                    <Add PerformanceCounter="\Process(??APP_WIN32_PROC??)\Handle Count" ReportAs="Process handle count" />
-                    ...
-                </Counters>
-      
-                PerformanceCounter must be either \CategoryName(InstanceName)\CounterName or \CategoryName\CounterName
-      
-                NOTE: performance counters configuration will be lost upon NuGet upgrade.
-      
-                The following placeholders are supported as InstanceName:
-                    ??APP_WIN32_PROC?? - instance name of the application process  for Win32 counters.
-                    ??APP_W3SVC_PROC?? - instance name of the application IIS worker process for IIS/ASP.NET counters.
-                    ??APP_CLR_PROC?? - instance name of the application CLR process for .NET counters.
-            -->
-        </Add>
-        <Add Type="Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse.QuickPulseTelemetryModule, Microsoft.AI.PerfCounterCollector"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.DeveloperModeWithDebuggerAttachedTelemetryModule, Microsoft.AI.WindowsServer"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.UnhandledExceptionTelemetryModule, Microsoft.AI.WindowsServer"/>
-        <Add Type="Microsoft.ApplicationInsights.WindowsServer.UnobservedExceptionTelemetryModule, Microsoft.AI.WindowsServer"/>
-        <!--<Add Type="Microsoft.ApplicationInsights.WindowsServer.FirstChanceExceptionStatisticsTelemetryModule, Microsoft.AI.WindowsServer"/>-->
-    </TelemetryModules>
-    <TelemetryChannel Type="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.ServerTelemetryChannel, Microsoft.AI.ServerTelemetryChannel"/>
-    <!-- 
-        Learn more about Application Insights configuration with ApplicationInsights.config here: 
-        http://go.microsoft.com/fwlink/?LinkID=513840
-    
-        Note: If not present, please add <InstrumentationKey>Your Key</InstrumentationKey> to the top of this file.
-    -->
-</ApplicationInsights>
