+++ toofz.NecroDancer.Leaderboards.Services.Common/Settings.cs
@@ -3,7 +3,5 @@
     public class Settings
     {
         public int UpdateInterval { get; set; } = 60;
-        public string ApiUserNameEnvVar { get; set; }
-        public string ApiPasswordEnvVar { get; set; }
     }
 }
