--- toofz.NecroDancer.Leaderboards.Services.Installer/CommonFiles.wxs
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Wix xmlns="http://schemas.microsoft.com/wix/2006/wi">
-    <Fragment>
-        <ComponentGroup Id="CommonFiles" Directory="APPLICATIONFOLDER" Source="$(var.toofz.NecroDancer.Leaderboards.Services.Common.TargetDir)">
-            <Component Id="EntityFramework.dll" Guid="{B73D7A27-2558-4E1C-8698-5BF5AEAEBEED}">
-                <File Id="EntityFramework.dll" KeyPath="yes" Name="EntityFramework.dll" />
-            </Component>
-            <Component Id="EntityFramework.SqlServer.dll" Guid="{7837067E-A453-4892-9A2C-ECED5D24B223}">
-                <File Id="EntityFramework.SqlServer.dll" KeyPath="yes" Name="EntityFramework.SqlServer.dll" />
-            </Component>
-            <Component Id="FluentValidation.dll" Guid="{5DA25A8A-1C8A-4FAC-A9AB-1BDE7E413F8A}">
-                <File Id="FluentValidation.dll" KeyPath="yes" Name="FluentValidation.dll" />
-            </Component>
-            <Component Id="log4net.dll" Guid="{A89F5FBF-CCC7-4CE9-8B6C-BC0E85881402}">
-                <File Id="log4net.dll" KeyPath="yes" Name="log4net.dll" />
-            </Component>
-            <Component Id="Microsoft.AI.Agent.Intercept.dll" Guid="{140E6B4C-4257-4E1A-AA50-1888F72D97F2}">
-                <File Id="Microsoft.AI.Agent.Intercept.dll" KeyPath="yes" Name="Microsoft.AI.Agent.Intercept.dll" />
-            </Component>
-            <Component Id="Microsoft.AI.DependencyCollector.dll" Guid="{476A8ADB-0404-4F3E-8EC9-C47C0FE992C0}">
-                <File Id="Microsoft.AI.DependencyCollector.dll" KeyPath="yes" Name="Microsoft.AI.DependencyCollector.dll" />
-            </Component>
-            <Component Id="Microsoft.AI.PerfCounterCollector.dll" Guid="{C2053A17-0B28-49DD-B551-79096B9D298E}">
-                <File Id="Microsoft.AI.PerfCounterCollector.dll" KeyPath="yes" Name="Microsoft.AI.PerfCounterCollector.dll" />
-            </Component>
-            <Component Id="Microsoft.AI.ServerTelemetryChannel.dll" Guid="{4C47A643-8BB8-42EA-851E-68627F08E0A8}">
-                <File Id="Microsoft.AI.ServerTelemetryChannel.dll" KeyPath="yes" Name="Microsoft.AI.ServerTelemetryChannel.dll" />
-            </Component>
-            <Component Id="Microsoft.AI.WindowsServer.dll" Guid="{A6615A8A-3DA8-47D5-BCB1-1DD7B3DC1A87}">
-                <File Id="Microsoft.AI.WindowsServer.dll" KeyPath="yes" Name="Microsoft.AI.WindowsServer.dll" />
-            </Component>
-            <Component Id="Microsoft.ApplicationInsights.dll" Guid="{B2D272C3-8E5B-46D3-B513-94CCBF4176B4}">
-                <File Id="Microsoft.ApplicationInsights.dll" KeyPath="yes" Name="Microsoft.ApplicationInsights.dll" />
-            </Component>
-            <Component Id="Microsoft.ApplicationInsights.Log4NetAppender.dll" Guid="{11FDB29E-747A-44F6-910E-9DB7FA4E266C}">
-                <File Id="Microsoft.ApplicationInsights.Log4NetAppender.dll" KeyPath="yes" Name="Microsoft.ApplicationInsights.Log4NetAppender.dll" />
-            </Component>
-            <Component Id="Microsoft.Azure.KeyVault.Core.dll" Guid="{DB37C87F-2455-4690-BB41-18942BA053FB}">
-                <File Id="Microsoft.Azure.KeyVault.Core.dll" KeyPath="yes" Name="Microsoft.Azure.KeyVault.Core.dll" />
-            </Component>
-            <Component Id="Microsoft.Data.Edm.dll" Guid="{C710FA2D-BC1F-4C0D-BECC-AE6B676A4C04}">
-                <File Id="Microsoft.Data.Edm.dll" KeyPath="yes" Name="Microsoft.Data.Edm.dll" />
-            </Component>
-            <Component Id="Microsoft.Data.OData.dll" Guid="{4243B38B-D54A-470E-975E-C3D13F4E0F8B}">
-                <File Id="Microsoft.Data.OData.dll" KeyPath="yes" Name="Microsoft.Data.OData.dll" />
-            </Component>
-            <Component Id="Microsoft.Data.Services.Client.dll" Guid="{262FFDAA-BF4A-41B6-979B-718A54441514}">
-                <File Id="Microsoft.Data.Services.Client.dll" KeyPath="yes" Name="Microsoft.Data.Services.Client.dll" />
-            </Component>
-            <Component Id="Microsoft.Practices.EnterpriseLibrary.TransientFaultHandling.dll" Guid="{9E1F05AB-F203-4643-9153-C11249E83829}">
-                <File Id="Microsoft.Practices.EnterpriseLibrary.TransientFaultHandling.dll" KeyPath="yes" Name="Microsoft.Practices.EnterpriseLibrary.TransientFaultHandling.dll" />
-            </Component>
-            <Component Id="Microsoft.WindowsAzure.Storage.dll" Guid="{7BDF6191-240E-4023-82E2-E89DC9C73F5E}">
-                <File Id="Microsoft.WindowsAzure.Storage.dll" KeyPath="yes" Name="Microsoft.WindowsAzure.Storage.dll" />
-            </Component>
-            <Component Id="Newtonsoft.Json.dll" Guid="{ECF4DFFA-1872-413A-8430-71AF83F8C882}">
-                <File Id="Newtonsoft.Json.dll" KeyPath="yes" Name="Newtonsoft.Json.dll" />
-            </Component>
-            <Component Id="protobuf_net.dll" Guid="{08BE8B4F-F214-464A-ACC6-9FC7DBD3D7F8}">
-                <File Id="protobuf_net.dll" KeyPath="yes" Name="protobuf-net.dll" />
-            </Component>
-            <Component Id="SqlBulkUpsert.dll" Guid="{1C5ECB00-9F0B-4A40-B6C1-CF984D97ECBC}">
-                <File Id="SqlBulkUpsert.dll" KeyPath="yes" Name="SqlBulkUpsert.dll" />
-            </Component>
-            <Component Id="SteamKit2.dll" Guid="{E6320479-541B-476D-8669-286493B5B39E}">
-                <File Id="SteamKit2.dll" KeyPath="yes" Name="SteamKit2.dll" />
-            </Component>
-            <Component Id="System.Net.Http.Formatting.dll" Guid="{DFA74BC9-122A-4694-94AB-A3157E72F47F}">
-                <File Id="System.Net.Http.Formatting.dll" KeyPath="yes" Name="System.Net.Http.Formatting.dll" />
-            </Component>
-            <Component Id="System.Spatial.dll" Guid="{E50CC2B4-32F8-4C23-852E-A3B5E0B7DFB9}">
-                <File Id="System.Spatial.dll" KeyPath="yes" Name="System.Spatial.dll" />
-            </Component>
-            <Component Id="System.Threading.Tasks.Dataflow.dll" Guid="{FBA3F574-C0F3-459F-A7CC-E99EC8CD961B}">
-                <File Id="System.Threading.Tasks.Dataflow.dll" KeyPath="yes" Name="System.Threading.Tasks.Dataflow.dll" />
-            </Component>
-            <Component Id="toofz.dll" Guid="{1C2D0AE3-713D-4A16-9BBF-BDC766D782DB}">
-                <File Id="toofz.dll" KeyPath="yes" Name="toofz.dll" />
-            </Component>
-            <Component Id="toofz.NecroDancer.dll" Guid="{A0BB783A-579D-4A7F-8800-FB09B658A8DB}">
-                <File Id="toofz.NecroDancer.dll" KeyPath="yes" Name="toofz.NecroDancer.dll" />
-            </Component>
-            <Component Id="toofz.NecroDancer.Leaderboards.dll" Guid="{278C283A-0F41-4F95-9B1A-11F8327D3B05}">
-                <File Id="toofz.NecroDancer.Leaderboards.dll" KeyPath="yes" Name="toofz.NecroDancer.Leaderboards.dll" />
-            </Component>
-            <Component Id="toofz.NecroDancer.Leaderboards.Services.Common.dll" Guid="{EABC2BD3-0646-4E34-8048-44E28725AAF8}">
-                <File Id="toofz.NecroDancer.Leaderboards.Services.Common.dll" KeyPath="yes" Name="toofz.NecroDancer.Leaderboards.Services.Common.dll" />
-            </Component>
-            <Component Id="ApplicationInsights.config" Guid="{C39D3B2E-2884-48C6-BB43-0F491B30B65F}">
-                <File Id="ApplicationInsights.config" KeyPath="yes" Name="ApplicationInsights.config" />
-            </Component>
-            <Component Id="daily.leaderboard.headers.min.json" Guid="{62948D22-B18B-4D7B-8D04-8C2BE446840F}">
-                <File Id="daily.leaderboard.headers.min.json" KeyPath="yes" Name="daily-leaderboard-headers.min.json" />
-            </Component>
-            <Component Id="leaderboard.headers.min.json" Guid="{CD9B018E-EDB7-45EB-B86D-6092C3D92907}">
-                <File Id="leaderboard.headers.min.json" KeyPath="yes" Name="leaderboard-headers.min.json" />
-            </Component>
-        </ComponentGroup>
-    </Fragment>
-</Wix>
