--- toofz.NecroDancer.Leaderboards.Services.Installer/toofz.NecroDancer.Leaderboards.Services.Installer.wixproj
@@ -1,84 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">x86</Platform>
-    <ProductVersion>3.10</ProductVersion>
-    <ProjectGuid>6d2c403a-153a-4093-b3a1-c5a54c594ae6</ProjectGuid>
-    <SchemaVersion>2.0</SchemaVersion>
-    <OutputName>toofz-leaderboard-services</OutputName>
-    <OutputType>Package</OutputType>
-    <WixTargetsPath Condition=" '$(WixTargetsPath)' == '' AND '$(MSBuildExtensionsPath32)' != '' ">$(MSBuildExtensionsPath32)\Microsoft\WiX\v3.x\Wix.targets</WixTargetsPath>
-    <WixTargetsPath Condition=" '$(WixTargetsPath)' == '' ">$(MSBuildExtensionsPath)\Microsoft\WiX\v3.x\Wix.targets</WixTargetsPath>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x86' ">
-    <OutputPath>bin\$(Configuration)\</OutputPath>
-    <IntermediateOutputPath>obj\$(Configuration)\</IntermediateOutputPath>
-    <DefineConstants>Debug</DefineConstants>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x86' ">
-    <OutputPath>bin\$(Configuration)\</OutputPath>
-    <IntermediateOutputPath>obj\$(Configuration)\</IntermediateOutputPath>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.Common\toofz.NecroDancer.Leaderboards.Services.Common.csproj">
-      <Name>toofz.NecroDancer.Leaderboards.Services.Common</Name>
-      <Project>{a80ace15-8a80-4912-aed0-385cc94b3974}</Project>
-      <Private>True</Private>
-      <DoNotHarvest>True</DoNotHarvest>
-      <RefProjectOutputGroups>Binaries;Content;Satellites</RefProjectOutputGroups>
-      <RefTargetDir>INSTALLFOLDER</RefTargetDir>
-    </ProjectReference>
-    <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.csproj">
-      <Name>toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate</Name>
-      <Project>{1678334c-de8c-4dc4-b493-4454bc209e18}</Project>
-      <Private>True</Private>
-      <DoNotHarvest>True</DoNotHarvest>
-      <RefProjectOutputGroups>Binaries;Content;Satellites</RefProjectOutputGroups>
-      <RefTargetDir>INSTALLFOLDER</RefTargetDir>
-    </ProjectReference>
-    <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.PlayerUpdate\toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.csproj">
-      <Name>toofz.NecroDancer.Leaderboards.Services.PlayerUpdate</Name>
-      <Project>{714c9b00-9392-4f2a-872c-3fe575a23d12}</Project>
-      <Private>True</Private>
-      <DoNotHarvest>True</DoNotHarvest>
-      <RefProjectOutputGroups>Binaries;Content;Satellites</RefProjectOutputGroups>
-      <RefTargetDir>INSTALLFOLDER</RefTargetDir>
-    </ProjectReference>
-    <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.ReplayUpdate\toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.csproj">
-      <Name>toofz.NecroDancer.Leaderboards.Services.ReplayUpdate</Name>
-      <Project>{a1dcec28-3283-4953-bd44-105dccbb04ce}</Project>
-      <Private>True</Private>
-      <DoNotHarvest>True</DoNotHarvest>
-      <RefProjectOutputGroups>Binaries;Content;Satellites</RefProjectOutputGroups>
-      <RefTargetDir>INSTALLFOLDER</RefTargetDir>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <WixExtension Include="WixUIExtension">
-      <HintPath>$(WixExtDir)\WixUIExtension.dll</HintPath>
-      <Name>WixUIExtension</Name>
-    </WixExtension>
-    <WixExtension Include="WixUtilExtension">
-      <HintPath>$(WixExtDir)\WixUtilExtension.dll</HintPath>
-      <Name>WixUtilExtension</Name>
-    </WixExtension>
-  </ItemGroup>
-  <Import Project="$(WixTargetsPath)" />
-  <ItemGroup>
-    <Compile Include="CommonFiles.wxs" />
-    <Compile Include="UnwelcomingFeatureTree.wxs" />
-    <Compile Include="LeaderboardService.wxs" />
-    <Compile Include="PlayerService.wxs" />
-    <Compile Include="Product.wxs" />
-    <Compile Include="ReplayService.wxs" />
-  </ItemGroup>
-  <Target Name="BeforeBuild">
-    <GetAssemblyIdentity AssemblyFiles="..\toofz.NecroDancer.Leaderboards.Services.Common\bin\$(Configuration)\toofz.NecroDancer.Leaderboards.Services.Common.dll">
-      <Output TaskParameter="Assemblies" ItemName="Assembly" />
-    </GetAssemblyIdentity>
-    <PropertyGroup>
-      <DefineConstants>Version=%(Assembly.Version)</DefineConstants>
-    </PropertyGroup>
-  </Target>
-</Project>
