+++ toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests/packages.config
@@ -2,7 +2,6 @@
 <packages>
   <package id="Castle.Core" version="4.1.1" targetFramework="net45" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
-  <package id="Microsoft.Tpl.Dataflow" version="4.5.24" targetFramework="net45" />
   <package id="Moq" version="4.7.99" targetFramework="net45" />
   <package id="MSTest.TestAdapter" version="1.1.18" targetFramework="net45" />
   <package id="MSTest.TestFramework" version="1.1.18" targetFramework="net45" />
