+++ toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests/toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{81D4C7F2-4DEA-4A5F-B818-55FCC25C350A}</ProjectGuid>
+    <ProjectGuid>{45E158FD-FCBB-409D-A09B-D54EDD758910}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>toofz.NecroDancer.Leaderboards.Tests</RootNamespace>
-    <AssemblyName>toofz.NecroDancer.Leaderboards.Tests</AssemblyName>
+    <RootNamespace>toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests</RootNamespace>
+    <AssemblyName>toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -71,55 +71,34 @@
     <Reference Include="System" />
     <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Net.Http" />
-    <Reference Include="System.Threading.Tasks.Dataflow, Version=4.5.24.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
-      <HintPath>..\packages\Microsoft.Tpl.Dataflow.4.5.24\lib\portable-net45+win8+wpa81\System.Threading.Tasks.Dataflow.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
+    <Reference Include="System.ServiceProcess" />
     <Reference Include="toofz.TestsShared, Version=1.1.0.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\toofz.TestsShared.1.1.0\lib\net45\toofz.TestsShared.dll</HintPath>
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.Common\toofz.NecroDancer.Leaderboards.Services.Common.csproj">
+      <Project>{a80ace15-8a80-4912-aed0-385cc94b3974}</Project>
+      <Name>toofz.NecroDancer.Leaderboards.Services.Common</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.csproj">
+      <Project>{1678334c-de8c-4dc4-b493-4454bc209e18}</Project>
+      <Name>toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate</Name>
+    </ProjectReference>
     <ProjectReference Include="..\toofz.NecroDancer.Leaderboards\toofz.NecroDancer.Leaderboards.csproj">
       <Project>{2b492909-63e4-4f96-898e-802ec5a37b0f}</Project>
       <Name>toofz.NecroDancer.Leaderboards</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Properties\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-      <SubType>Designer</SubType>
-    </EmbeddedResource>
-    <Compile Include="Properties\Resources.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
-    <None Include="Resources\2_9_2014_score191_zone1_level2" />
-    <None Include="Resources\GetPlayerSummaries.json" />
-    <None Include="Resources\GetUgcFileDetails.json" />
-    <Compile Include="Steam\ClientApi\ProgressDebugNetworkListenerTests.cs" />
-    <Compile Include="Steam\ClientApi\SteamClientApiClientTests.cs" />
-    <Compile Include="Steam\ClientApi\SteamClientApiTransientErrorDetectionStrategyTests.cs" />
-    <Compile Include="Steam\WebApi\ISteamRemoteStorage\UgcFileDetailsTests.cs" />
-    <Compile Include="Steam\WebApi\ISteamUser\PlayerSummariesTests.cs" />
-    <Compile Include="Steam\WebApi\SteamWebApiClientTests.cs" />
-    <Compile Include="ApiClientTests.cs" />
-    <Compile Include="ApiExceptionTests.cs" />
-    <Compile Include="Constants.cs" />
-    <Compile Include="EnsureSuccessHandlerTests.cs" />
-    <Compile Include="HttpRequestStatusHandlerTests.cs" />
-    <Compile Include="ITargetBlockExtensionsTests.cs" />
-    <Compile Include="OAuth2HandlerTests.cs" />
-    <Compile Include="StreamPipelineTests.cs" />
-    <Compile Include="UgcHttpClientTests.cs" />
-    <None Include="App.config">
-      <SubType>Designer</SubType>
-    </None>
-    <None Include="packages.config">
-      <SubType>Designer</SubType>
-    </None>
+    <Folder Include="Properties\" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="App.config" />
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="WorkerRoleTests.cs" />
   </ItemGroup>
   <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
