+++ toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate/Properties/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("toofz Leaderboard Service")]
@@ -8,5 +9,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("Copyright © Leonard Thieu 2015")]
 [assembly: ComVisible(false)]
 [assembly: AssemblyVersion("1.1.*")]
+
+[assembly: InternalsVisibleTo("toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests")]
 
 [assembly: log4net.Config.XmlConfigurator(Watch = true, ConfigFile = "leaderboard-log.config")]
