+++ toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate/packages.config
@@ -9,6 +9,8 @@
   <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.4.1" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.4.1" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.4.0" targetFramework="net45" />
+  <package id="protobuf-net" version="2.0.0.668" targetFramework="net45" />
+  <package id="SteamKit2" version="1.8.3" targetFramework="net45" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.4.0" targetFramework="net45" />
   <package id="toofz" version="2.1.0" targetFramework="net45" />
 </packages>
