+++ toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests/packages.config
@@ -2,14 +2,11 @@
 <packages>
   <package id="Castle.Core" version="4.1.1" targetFramework="net45" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
-  <package id="Microsoft.Tpl.Dataflow" version="4.5.24" targetFramework="net45" />
   <package id="Moq" version="4.7.99" targetFramework="net45" />
   <package id="MSTest.TestAdapter" version="1.1.18" targetFramework="net45" />
   <package id="MSTest.TestFramework" version="1.1.18" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
-  <package id="protobuf-net" version="2.0.0.668" targetFramework="net45" />
   <package id="RichardSzalay.MockHttp" version="1.5.0" targetFramework="net45" />
-  <package id="SteamKit2" version="1.8.3" targetFramework="net45" />
   <package id="System.Net.Http" version="4.3.2" targetFramework="net45" />
   <package id="toofz.TestsShared" version="1.1.0" targetFramework="net45" />
 </packages>
