+++ toofz.NecroDancer.Leaderboards.Services.PlayerUpdate/packages.config
@@ -8,5 +8,8 @@
   <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.4.1" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.4.1" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.4.0" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.4.0" targetFramework="net45" />
+  <package id="toofz" version="2.1.0" targetFramework="net45" />
 </packages>
