+++ toofz.NecroDancer.Leaderboards.Services.PlayerUpdate/player-settings.json
@@ -1,6 +1,4 @@
 ﻿{
   "UpdateInterval": 30,
-  "ApiUserNameEnvVar": "PlayersUserName",
-  "ApiPasswordEnvVar": "PlayersPassword",
   "PlayersPerUpdate": 500
 }
