+++ toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.Tests/packages.config
@@ -1,0 +1,22 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Castle.Core" version="4.1.1" targetFramework="net45" />
+  <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
+  <package id="Microsoft.Azure.KeyVault.Core" version="2.0.4" targetFramework="net45" />
+  <package id="Microsoft.Data.Edm" version="5.8.2" targetFramework="net45" />
+  <package id="Microsoft.Data.OData" version="5.8.2" targetFramework="net45" />
+  <package id="Microsoft.Data.Services.Client" version="5.8.2" targetFramework="net45" />
+  <package id="Moq" version="4.7.99" targetFramework="net45" />
+  <package id="MSTest.TestAdapter" version="1.1.18" targetFramework="net45" />
+  <package id="MSTest.TestFramework" version="1.1.18" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
+  <package id="RichardSzalay.MockHttp" version="1.5.0" targetFramework="net45" />
+  <package id="System.ComponentModel.EventBasedAsync" version="4.3.0" targetFramework="net45" />
+  <package id="System.Dynamic.Runtime" version="4.3.0" targetFramework="net45" />
+  <package id="System.Linq.Queryable" version="4.3.0" targetFramework="net45" />
+  <package id="System.Net.Http" version="4.3.2" targetFramework="net45" />
+  <package id="System.Net.Requests" version="4.3.0" targetFramework="net45" />
+  <package id="System.Spatial" version="5.8.2" targetFramework="net45" />
+  <package id="toofz.TestsShared" version="1.1.0" targetFramework="net45" />
+  <package id="WindowsAzure.Storage" version="8.2.1" targetFramework="net45" />
+</packages>
