+++ toofz.NecroDancer.Leaderboards.Services.ReplayUpdate/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="FluentValidation" version="7.1.1" targetFramework="net45" />
   <package id="log4net" version="2.0.8" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights" version="2.4.0" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.4.0" targetFramework="net45" />
@@ -8,11 +9,11 @@
   <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.4.1" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.4.1" targetFramework="net45" />
   <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.4.0" targetFramework="net45" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net45" />
   <package id="Microsoft.Azure.KeyVault.Core" version="2.0.4" targetFramework="net45" />
   <package id="Microsoft.Data.Edm" version="5.8.2" targetFramework="net45" />
   <package id="Microsoft.Data.OData" version="5.8.2" targetFramework="net45" />
   <package id="Microsoft.Data.Services.Client" version="5.8.2" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.3" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
   <package id="System.ComponentModel.EventBasedAsync" version="4.3.0" targetFramework="net45" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.4.0" targetFramework="net45" />
@@ -21,5 +22,6 @@
   <package id="System.Net.Requests" version="4.3.0" targetFramework="net45" />
   <package id="System.Spatial" version="5.8.2" targetFramework="net45" />
   <package id="toofz" version="2.1.0" targetFramework="net45" />
+  <package id="toofz.NecroDancer" version="2.0.0" targetFramework="net45" />
   <package id="WindowsAzure.Storage" version="8.2.1" targetFramework="net45" />
 </packages>
