+++ toofz.NecroDancer.Leaderboards.Services.ReplayUpdate/replay-settings.json
@@ -1,6 +1,4 @@
 ﻿{
   "UpdateInterval": 60,
-  "ApiUserNameEnvVar": "ReplaysUserName",
-  "ApiPasswordEnvVar": "ReplaysPassword",
   "ReplaysPerUpdate": 60
 }
