+++ toofz.NecroDancer.Web.sln
@@ -7,8 +7,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboards.Services.Common", "toofz.NecroDancer.Leaderboards.Services.Common\toofz.NecroDancer.Leaderboards.Services.Common.csproj", "{A80ACE15-8A80-4912-AED0-385CC94B3974}"
 EndProject
-Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "toofz.NecroDancer.Leaderboards.Services.Installer", "toofz.NecroDancer.Leaderboards.Services.Installer\toofz.NecroDancer.Leaderboards.Services.Installer.wixproj", "{6D2C403A-153A-4093-B3A1-C5A54C594AE6}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate", "toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.csproj", "{1678334C-DE8C-4DC4-B493-4454BC209E18}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboards.Services.PlayerUpdate", "toofz.NecroDancer.Leaderboards.Services.PlayerUpdate\toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.csproj", "{714C9B00-9392-4F2A-872C-3FE575A23D12}"
@@ -33,6 +31,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Web.Api.Tests", "toofz.NecroDancer.Web.Api.Tests\toofz.NecroDancer.Web.Api.Tests.csproj", "{AF85C3C1-6A21-43E4-B9B6-E0D2707D9A9E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests", "toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests\toofz.NecroDancer.Leaderboards.Services.LeaderboardUpdate.Tests.csproj", "{45E158FD-FCBB-409D-A09B-D54EDD758910}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests", "toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests\toofz.NecroDancer.Leaderboards.Services.PlayerUpdate.Tests.csproj", "{3F602AB5-05E8-4A2D-B84A-F76EE07BE5EE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.Tests", "toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.Tests\toofz.NecroDancer.Leaderboards.Services.ReplayUpdate.Tests.csproj", "{100F18F7-93AB-4493-92B0-FF5541373FAC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,9 +51,6 @@ Global
 		{A80ACE15-8A80-4912-AED0-385CC94B3974}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A80ACE15-8A80-4912-AED0-385CC94B3974}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A80ACE15-8A80-4912-AED0-385CC94B3974}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6D2C403A-153A-4093-B3A1-C5A54C594AE6}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{6D2C403A-153A-4093-B3A1-C5A54C594AE6}.Release|Any CPU.ActiveCfg = Release|x86
-		{6D2C403A-153A-4093-B3A1-C5A54C594AE6}.Release|Any CPU.Build.0 = Release|x86
 		{1678334C-DE8C-4DC4-B493-4454BC209E18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1678334C-DE8C-4DC4-B493-4454BC209E18}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1678334C-DE8C-4DC4-B493-4454BC209E18}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -74,6 +75,18 @@ Global
 		{AF85C3C1-6A21-43E4-B9B6-E0D2707D9A9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AF85C3C1-6A21-43E4-B9B6-E0D2707D9A9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AF85C3C1-6A21-43E4-B9B6-E0D2707D9A9E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45E158FD-FCBB-409D-A09B-D54EDD758910}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45E158FD-FCBB-409D-A09B-D54EDD758910}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45E158FD-FCBB-409D-A09B-D54EDD758910}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45E158FD-FCBB-409D-A09B-D54EDD758910}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3F602AB5-05E8-4A2D-B84A-F76EE07BE5EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3F602AB5-05E8-4A2D-B84A-F76EE07BE5EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3F602AB5-05E8-4A2D-B84A-F76EE07BE5EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3F602AB5-05E8-4A2D-B84A-F76EE07BE5EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{100F18F7-93AB-4493-92B0-FF5541373FAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{100F18F7-93AB-4493-92B0-FF5541373FAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{100F18F7-93AB-4493-92B0-FF5541373FAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{100F18F7-93AB-4493-92B0-FF5541373FAC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -81,13 +94,15 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{2B492909-63E4-4F96-898E-802EC5A37B0F} = {DF291D24-732A-49D4-9F21-33712811F79D}
 		{A80ACE15-8A80-4912-AED0-385CC94B3974} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
-		{6D2C403A-153A-4093-B3A1-C5A54C594AE6} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
 		{1678334C-DE8C-4DC4-B493-4454BC209E18} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
 		{714C9B00-9392-4F2A-872C-3FE575A23D12} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
 		{A1DCEC28-3283-4953-BD44-105DCCBB04CE} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
 		{538C3CD9-BE40-4CD6-AEC8-4B7C07EC4CE7} = {E1862536-1291-47AB-8DF8-65C3156CB8C8}
 		{81D4C7F2-4DEA-4A5F-B818-55FCC25C350A} = {DF291D24-732A-49D4-9F21-33712811F79D}
 		{AF85C3C1-6A21-43E4-B9B6-E0D2707D9A9E} = {E1862536-1291-47AB-8DF8-65C3156CB8C8}
+		{45E158FD-FCBB-409D-A09B-D54EDD758910} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
+		{3F602AB5-05E8-4A2D-B84A-F76EE07BE5EE} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
+		{100F18F7-93AB-4493-92B0-FF5541373FAC} = {CC7FC22B-EAC1-48BE-8149-9BE3DC096509}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		EnterpriseLibraryConfigurationToolBinariesPathV6 = packages\EnterpriseLibrary.TransientFaultHandling.6.0.1304.0\lib\portable-net45+win+wp8
